pyproject.toml
[tool.poetry]
name = "machine-learning"
version = "0.1.0"
description = ""
authors = ["ferox <malov.egor.alexandrovich@gmail.com >"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
matplotlib = "^3.8.0"
black = "^23.9.1"
pandas = "^2.1.0"
tqdm = "^4.66.1"
python-dotenv = "^1.0.0"
dash = "^2.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-versions = ["py311"]
line-length = 120
skip-string-normalization = true
